+++ js/modules/manifest.js
@@ -1,4 +1,4 @@
-const CACHE_BUST = 'v=32';
+const CACHE_BUST = 'v=33';
 
 export const vendorScriptsCritical = [
   'https://cdnjs.cloudflare.com/ajax/libs/three.js/r128/three.min.js',